requirements.txt
hydra-core
beautifulsoup4
tqdm
loguru